refactor: refactor metaPanel, divide files in metaPanel
