+++ frontend/package.json
@@ -11,7 +11,6 @@
     },
     "dependencies": {
         "@craco/craco": "^6.3.0",
-        "@fortawesome/fontawesome-svg-core": "^1.2.36",
         "@reduxjs/toolkit": "^1.6.0",
         "autoprefixer": "^9",
         "axios": "^0.21.1",
