+++ frontend/src/components/server/metaPanel/redux/metaPanel.slice.ts
@@ -1,8 +1,10 @@
 import { createSlice } from '@reduxjs/toolkit'
-import { UserInfo } from 'components/auth/redux/auth.slice'
 import { RootState } from 'redux/store'
+
+import { UserInfo } from 'components/auth/redux/auth.slice'
+import { ChannelInfo } from 'components/server/redux/channels/channels.slice'
+
 import { Undefinable, WithPayload } from 'types/commonType'
-import { ChannelInfo } from './channels/channels.slice'
 
 interface MetaPanelInitialState {
     isOpen: boolean
